+++ lib/std/os/bits/freebsd.zig
@@ -133,7 +133,7 @@ pub const dirent = extern struct {
     d_pad1: u16,
     d_name: [256]u8,
 
-    pub fn reclen(self: dirent) usize {
+    pub fn reclen(self: dirent) u16 {
         return self.d_reclen;
     }
 };
